moved cleanText function to helper folder
