+++ contact/index.html
@@ -23,7 +23,7 @@
   <body id="body" class="grey-section">
     <div class="page-content">
       <div>
-        <header>
+        <header id="header-container">
           <nav class="header-nav">
             <div class="nav-links-left">
               <a href="../contact/index.html" class="nav-link current-page" title="contact me">Contact</a>
@@ -171,8 +171,8 @@
               </div>
             </div>
             <div class="top-btn-container">
-              <div class="top-btn">
-                <a href="#" title="to top"><span class="material-symbols-sharp size-32 arrow-up"> keyboard_arrow_up</span></a>
+              <div class="top-btn" id="top-btn">
+                <button type="button" title="to top"><span class="material-symbols-sharp size-32 arrow-up">keyboard_arrow_up</span></button>
               </div>
               <p>top</p>
             </div>
