+++ js/ui/displayCarouselCardsByDate.mjs
@@ -1,6 +1,6 @@
 import { displayDynamicDate } from "../components/displayDynamicDate.mjs";
 import { createPostImage } from "../components/createPostImage.mjs";
-import { cleanText } from "../components/cleanText.mjs";
+import { cleanText } from "../helper/cleanText.mjs";
 
 export const displayCarouselCardsByDate = (container, posts, numberOfPosts) => {
   // exit the loading indicator when loaded
