+++ js/views/about.mjs
@@ -1,5 +1,11 @@
 import { toggleMobileNav } from "../helper/toggleMobileNav.mjs";
-toggleMobileNav();
-
 import { displayCarousel } from "../handler/displayCarousel.mjs";
+import { scrollToSectionByClick } from "../helper/scrollToSectionByClick.mjs";
+import { toTopButtonWrapper, headerContainer } from "../data/constants.mjs";
+
+// toggle mobile nav
+toggleMobileNav();
+// display posts carousel
 displayCarousel();
+// scroll to header by "to top button"
+scrollToSectionByClick(toTopButtonWrapper, headerContainer);
