+++ js/views/articles.mjs
@@ -1,6 +1,8 @@
 import { toggleMobileNav } from "../helper/toggleMobileNav.mjs";
 import { handlePosts } from "../handler/handlePosts.mjs";
 import { postWrapper } from "../data/constants.mjs";
+import { scrollToSectionByClick } from "../helper/scrollToSectionByClick.mjs";
+import { toTopButtonWrapper, headerContainer } from "../data/constants.mjs";
 
 // trying to increase loading speeds in terms of Lighthouse statistics
 import { fetchAllPosts } from "../api/fetchAllPosts.mjs";
@@ -12,3 +14,6 @@ document.addEventListener("DOMContentLoaded", () => {
 toggleMobileNav();
 // fetching and displaying posts
 handlePosts(postWrapper);
+
+// scroll to header by "to top button"
+scrollToSectionByClick(toTopButtonWrapper, headerContainer);
